PATCH 5.10 documentation
A bunch of changes to the documentation.  They range from fixing minor
typos to changing outright lies, and things in between.

Much of it has to do with Unicode, and includes mostly things that I 
found out the hard way.

There are a few comment changes, but almost all is in pods or used in 
generating pods.

p5p-msgid: <49483312.80804@khwilliamson.com>